Refactor PrivilegesLoader (#9768)
