+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/metadata/auth/builder/loader/dialect/MySQLPrivilegeLoader.java
@@ -17,21 +17,25 @@
 
 package org.apache.shardingsphere.infra.metadata.auth.builder.loader.dialect;
 
+import com.google.common.base.Joiner;
 import org.apache.shardingsphere.infra.metadata.auth.builder.loader.PrivilegeLoader;
 import org.apache.shardingsphere.infra.metadata.auth.model.privilege.PrivilegeType;
 import org.apache.shardingsphere.infra.metadata.auth.model.privilege.ShardingSpherePrivilege;
 import org.apache.shardingsphere.infra.metadata.auth.model.privilege.database.SchemaPrivilege;
 import org.apache.shardingsphere.infra.metadata.auth.model.privilege.database.TablePrivilege;
+import org.apache.shardingsphere.infra.metadata.auth.model.user.Grantee;
 import org.apache.shardingsphere.infra.metadata.auth.model.user.ShardingSphereUser;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.LinkedHashMap;
 import java.util.LinkedList;
+import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -41,25 +45,140 @@ import java.util.stream.Collectors;
 public final class MySQLPrivilegeLoader implements PrivilegeLoader {
     
     @Override
-    public Optional<ShardingSpherePrivilege> load(final ShardingSphereUser user, final DataSource dataSource) throws SQLException {
-        ShardingSpherePrivilege result = new ShardingSpherePrivilege();
-        fillGlobalPrivilege(result, dataSource, user);
-        fillSchemaPrivilege(result, dataSource, user);
-        fillTablePrivilege(result, dataSource, user);
-        return Optional.of(result);
+    public Map<ShardingSphereUser, ShardingSpherePrivilege> load(final Collection<ShardingSphereUser> users, final DataSource dataSource) throws SQLException {
+        Map<ShardingSphereUser, ShardingSpherePrivilege> result = new LinkedHashMap<>();
+        fillGlobalPrivilege(result, dataSource, users);
+        fillSchemaPrivilege(result, dataSource, users);
+        fillTablePrivilege(result, dataSource, users);
+        return result;
     }
     
-    private void fillGlobalPrivilege(final ShardingSpherePrivilege privilege, final DataSource dataSource, final ShardingSphereUser user) throws SQLException {
+    private void fillGlobalPrivilege(final Map<ShardingSphereUser, ShardingSpherePrivilege> privileges, final DataSource dataSource, final Collection<ShardingSphereUser> users) throws SQLException {
         try (Connection connection = dataSource.getConnection()) {
-            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM mysql.user WHERE user=? AND host=?");
-            preparedStatement.setString(1, user.getGrantee().getUsername());
-            preparedStatement.setString(2, user.getGrantee().getHostname());
-            try (ResultSet resultSet = preparedStatement.executeQuery()) {
-                if (resultSet.first()) {
-                    privilege.getAdministrativePrivilege().getPrivileges().addAll(loadAdministrativePrivileges(resultSet));
-                    privilege.getDatabasePrivilege().getGlobalPrivileges().addAll(loadDatabaseGlobalPrivileges(resultSet));
+            Statement statement = connection.createStatement();
+            try (ResultSet resultSet = statement.executeQuery(getGlobalPrivilegeSQL(users))) {
+                while (resultSet.next()) {
+                    fillGlobalPrivilege(privileges, resultSet);
                 }
             }
+        }
+    }
+    
+    private void fillGlobalPrivilege(final Map<ShardingSphereUser, ShardingSpherePrivilege> privileges, final ResultSet resultSet) throws SQLException {
+        Optional<ShardingSphereUser> user = getShardingSphereUser(privileges, resultSet);
+        if (user.isPresent()) {
+            privileges.get(user.get()).getAdministrativePrivilege().getPrivileges().addAll(loadAdministrativePrivileges(resultSet));
+            privileges.get(user.get()).getDatabasePrivilege().getGlobalPrivileges().addAll(loadDatabaseGlobalPrivileges(resultSet));
+        }
+    }
+    
+    private void fillSchemaPrivilege(final Map<ShardingSphereUser, ShardingSpherePrivilege> privileges, final DataSource dataSource, final Collection<ShardingSphereUser> users) throws SQLException {
+        try (Connection connection = dataSource.getConnection()) {
+            Statement statement = connection.createStatement();
+            try (ResultSet resultSet = statement.executeQuery(getSchemaPrivilegeSQL(users))) {
+                while (resultSet.next()) {
+                    fillSchemaPrivilege(privileges, resultSet);
+                }
+            }
+        }
+    }
+    
+    private void fillSchemaPrivilege(final Map<ShardingSphereUser, ShardingSpherePrivilege> privileges, final ResultSet resultSet) throws SQLException {
+        Optional<ShardingSphereUser> user = getShardingSphereUser(privileges, resultSet);
+        if (user.isPresent()) {
+            String db = resultSet.getString("Db");
+            SchemaPrivilege schemaPrivilege = new SchemaPrivilege(db);
+            schemaPrivilege.getGlobalPrivileges().addAll(loadDatabaseGlobalPrivileges(resultSet));
+            privileges.get(user.get()).getDatabasePrivilege().getSpecificPrivileges().put(db, schemaPrivilege);
+        }
+    }
+    
+    private void fillTablePrivilege(final Map<ShardingSphereUser, ShardingSpherePrivilege> privileges, final DataSource dataSource, final Collection<ShardingSphereUser> users) throws SQLException {
+        try (Connection connection = dataSource.getConnection()) {
+            Statement statement = connection.createStatement();
+            try (ResultSet resultSet = statement.executeQuery(getTablePrivilegeSQL(users))) {
+                while (resultSet.next()) {
+                    fillTablePrivilege(privileges, resultSet);
+                }
+            }
+        }
+    }
+    
+    private void fillTablePrivilege(final Map<ShardingSphereUser, ShardingSpherePrivilege> privileges, final ResultSet resultSet) throws SQLException {
+        Optional<ShardingSphereUser> user = getShardingSphereUser(privileges, resultSet);
+        if (user.isPresent()) {
+            String db = resultSet.getString("Db");
+            String tableName = resultSet.getString("Table_name");
+            String[] tablePrivileges = (String[]) resultSet.getArray("Table_priv").getArray();
+            TablePrivilege tablePrivilege = new TablePrivilege(tableName, getPrivileges(tablePrivileges));
+            ShardingSpherePrivilege privilege = privileges.get(user.get());
+            if (!privilege.getDatabasePrivilege().getSpecificPrivileges().containsKey(db)) {
+                privilege.getDatabasePrivilege().getSpecificPrivileges().put(db, new SchemaPrivilege(db));
+            }
+            privilege.getDatabasePrivilege().getSpecificPrivileges().get(db).getSpecificPrivileges().put(tableName, tablePrivilege);
+        }
+    }
+    
+    private String getGlobalPrivilegeSQL(final Collection<ShardingSphereUser> users) {
+        Collection<String> result = new LinkedList<>();
+        StringBuilder builder = new StringBuilder("SELECT * FROM mysql.user WHERE (user, host) in ( ");
+        users.forEach(each -> builder.append("(").append(each.getGrantee().getUsername()).append(", ").append(each.getGrantee().getHostname()).append(")"));
+        return builder.append(Joiner.on(", ").join(result)).append(" )").toString();
+    }
+    
+    private String getSchemaPrivilegeSQL(final Collection<ShardingSphereUser> users) {
+        Collection<String> result = new LinkedList<>();
+        StringBuilder builder = new StringBuilder("SELECT * FROM mysql.db WHERE (user, host) in ( ");
+        users.forEach(each -> builder.append("(").append(each.getGrantee().getUsername()).append(", ").append(each.getGrantee().getHostname()).append(")"));
+        return builder.append(Joiner.on(", ").join(result)).append(" )").toString();
+    }
+    
+    private String getTablePrivilegeSQL(final Collection<ShardingSphereUser> users) {
+        Collection<String> result = new LinkedList<>();
+        StringBuilder builder = new StringBuilder("SELECT Db, Table_name, Table_priv FROM mysql.tables_priv WHERE (user, host) in ( ");
+        users.forEach(each -> builder.append("(").append(each.getGrantee().getUsername()).append(", ").append(each.getGrantee().getHostname()).append(")"));
+        return builder.append(Joiner.on(", ").join(result)).append(" )").toString();
+    }
+    
+    private Optional<ShardingSphereUser> getShardingSphereUser(final Map<ShardingSphereUser, ShardingSpherePrivilege> privileges, final ResultSet resultSet) throws SQLException {
+        Grantee grantee = new Grantee(resultSet.getString("user"), resultSet.getString("host"));
+        return privileges.keySet().stream().filter(each -> each.getGrantee().equals(grantee)).findFirst();
+    }
+    
+    private Collection<PrivilegeType> getPrivileges(final String[] privileges) {
+        return Arrays.stream(privileges).map(this::getPrivilegeType).collect(Collectors.toSet());
+    }
+    
+    private PrivilegeType getPrivilegeType(final String privilege) {
+        switch (privilege) {
+            case "Select":
+                return PrivilegeType.SELECT;
+            case "Insert":
+                return PrivilegeType.INSERT;
+            case "Update":
+                return PrivilegeType.UPDATE;
+            case "Delete":
+                return PrivilegeType.DELETE;
+            case "Create":
+                return PrivilegeType.CREATE;
+            case "Alter":
+                return PrivilegeType.ALTER;
+            case "Drop":
+                return PrivilegeType.DROP;
+            case "Grant":
+                return PrivilegeType.GRANT;
+            case "Index":
+                return PrivilegeType.INDEX;
+            case "References":
+                return PrivilegeType.REFERENCES;
+            case "Create View":
+                return PrivilegeType.CREATE_VIEW;
+            case "Show view":
+                return PrivilegeType.SHOW_VIEW;
+            case "Trigger":
+                return PrivilegeType.TRIGGER;
+            default:
+                throw new UnsupportedOperationException(privilege);
         }
     }
     
@@ -105,79 +224,6 @@ public final class MySQLPrivilegeLoader implements PrivilegeLoader {
     private void addToPrivilegeTypesIfPresent(final boolean hasPrivilege, final PrivilegeType privilegeType, final Collection<PrivilegeType> target) {
         if (hasPrivilege) {
             target.add(privilegeType);
-        }
-    }
-    
-    private void fillSchemaPrivilege(final ShardingSpherePrivilege privilege, final DataSource dataSource, final ShardingSphereUser user) throws SQLException {
-        try (Connection connection = dataSource.getConnection()) {
-            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM mysql.db WHERE user=? AND host=?");
-            preparedStatement.setString(1, user.getGrantee().getUsername());
-            preparedStatement.setString(2, user.getGrantee().getHostname());
-            try (ResultSet resultSet = preparedStatement.executeQuery()) {
-                while (resultSet.first()) {
-                    String db = resultSet.getString("Db");
-                    SchemaPrivilege schemaPrivilege = new SchemaPrivilege(db);
-                    schemaPrivilege.getGlobalPrivileges().addAll(loadDatabaseGlobalPrivileges(resultSet));
-                    privilege.getDatabasePrivilege().getSpecificPrivileges().put(db, schemaPrivilege);
-                }
-            }
-        }
-    }
-    
-    private void fillTablePrivilege(final ShardingSpherePrivilege privilege, final DataSource dataSource, final ShardingSphereUser user) throws SQLException {
-        try (Connection connection = dataSource.getConnection()) {
-            PreparedStatement preparedStatement = connection.prepareStatement("SELECT Db, Table_name, Table_priv FROM mysql.tables_priv WHERE user=? AND host=?");
-            preparedStatement.setString(1, user.getGrantee().getUsername());
-            preparedStatement.setString(2, user.getGrantee().getHostname());
-            try (ResultSet resultSet = preparedStatement.executeQuery()) {
-                while (resultSet.next()) {
-                    String db = resultSet.getString("Db");
-                    String tableName = resultSet.getString("Table_name");
-                    String[] tablePrivileges = (String[]) resultSet.getArray("Table_priv").getArray();
-                    TablePrivilege tablePrivilege = new TablePrivilege(tableName, getPrivileges(tablePrivileges));
-                    if (!privilege.getDatabasePrivilege().getSpecificPrivileges().containsKey(db)) {
-                        privilege.getDatabasePrivilege().getSpecificPrivileges().put(db, new SchemaPrivilege(db));
-                    }
-                    privilege.getDatabasePrivilege().getSpecificPrivileges().get(db).getSpecificPrivileges().put(tableName, tablePrivilege);
-                }
-            }
-        }
-    }
-    
-    private Collection<PrivilegeType> getPrivileges(final String[] privileges) {
-        return Arrays.stream(privileges).map(this::getPrivilegeType).collect(Collectors.toSet());
-    }
-    
-    private PrivilegeType getPrivilegeType(final String privilege) {
-        switch (privilege) {
-            case "Select":
-                return PrivilegeType.SELECT;
-            case "Insert":
-                return PrivilegeType.INSERT;
-            case "Update":
-                return PrivilegeType.UPDATE;
-            case "Delete":
-                return PrivilegeType.DELETE;
-            case "Create":
-                return PrivilegeType.CREATE;
-            case "Alter":
-                return PrivilegeType.ALTER;
-            case "Drop":
-                return PrivilegeType.DROP;
-            case "Grant":
-                return PrivilegeType.GRANT;
-            case "Index":
-                return PrivilegeType.INDEX;
-            case "References":
-                return PrivilegeType.REFERENCES;
-            case "Create View":
-                return PrivilegeType.CREATE_VIEW;
-            case "Show view":
-                return PrivilegeType.SHOW_VIEW;
-            case "Trigger":
-                return PrivilegeType.TRIGGER;
-            default:
-                throw new UnsupportedOperationException(privilege);
         }
     }
